+++ docs/changelog.rst
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`_.
 
+Unreleased
+----------
+
+Added
+#####
+- Clocking schemes:
+    - SRS
+
+
 v0.6.7 - 2025-01-20
 -------------------